Removed data_structures and put the code in __init__
